+++ src/test/java/menu/MenuTest.java
@@ -1,6 +1,6 @@
 package menu;
 
-import menu.domain.Coach;
+import menu.domain.coach.Coach;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
